+++ projects/simple-demo/src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
 export const environment = {
     production: true,
-    backend: 'http://127.0.0.1:8080/db/ping-pong',
+    backend: 'http://127.0.0.1:8081/db/site',
 };